Rename to ProcessingImagesRedux
