+++ App/Redux/ProcessingImagesRedux.ts
